+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- refactor: new type StarknetVersion
 - refactor: update starknet-rs with Felt
 - fix(rpc): fixed block not found error on get_class method
 - fix (rpc): get_transaction_status